feat: make mobile navbar
Started working on my portfolio. At first thought to make a white theme, but when saw how the dark theme looks and checked some templates, decided to settle in with that one. I develop mobile first, because it is a cleaner way to do it in my opinion and right now have only added a navbar, which is not yet responsive