feat: add main grid layout
Okay, so I think the main layout is kind of there, especially for the phones, I think it will work really well. But now I just have to figure out how to make it look good on tablets and PCs. Work in progress. Also, gotta change the images, so the next commit will probably be just that