feat: finish header
Okay, now I think the header is finished. It also had a little shenanigan where on the edges the underlying layer of a page, or simply the content under the header, could be seen better than more to the center, but that wsa due to a backdrop-blur set to XL, which caused it to blur more on the center and less on the edges, I think. A simple fix to it was to set the property to SM, and now its all good. We now have two menus - one for each button and they both open up on button press and stay visible on focus. Very nice!